fix(deletion): prevent external side-effect replay (#758)
Commit the deletion batch before sending workflow error notifications and treat a post-refresh Keycloak NotFound as an idempotent outcome.
