+++ src/test/java/de/caritas/cob/userservice/api/workflow/delete/scheduler/DeleteUserAnonymousSchedulerIT.java
@@ -2,8 +2,11 @@ package de.caritas.cob.userservice.api.workflow.delete.scheduler;
 
 import static de.caritas.cob.userservice.api.testHelper.TestConstants.CONSULTING_TYPE_ID_AIDS;
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.when;
 
 import de.caritas.cob.userservice.api.adapters.matrix.MatrixSynapseService;
@@ -21,6 +24,7 @@ import de.caritas.cob.userservice.api.testConfig.ApiControllerTestConfig;
 import de.caritas.cob.userservice.api.testConfig.ConsultingTypeManagerTestConfig;
 import de.caritas.cob.userservice.api.testConfig.KeycloakTestConfig;
 import de.caritas.cob.userservice.api.testConfig.TestAgencyControllerApi;
+import de.caritas.cob.userservice.api.workflow.delete.service.WorkflowErrorMailService;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -61,6 +65,8 @@ class DeleteUserAnonymousSchedulerIT {
   @MockitoBean AgencyServiceApiControllerFactory agencyServiceApiControllerFactory;
 
   @MockitoBean MatrixSynapseService matrixSynapseService;
+
+  @MockitoBean WorkflowErrorMailService workflowErrorMailService;
 
   private Session currentSession;
 
@@ -123,6 +129,21 @@ class DeleteUserAnonymousSchedulerIT {
     prepareCurrentSessionForDeletion();
 
     deleteUserAnonymousScheduler.performDeletionWorkflow();
+
+    assertSessionAndUserDoNotExistInDatabase(
+        currentSession.getId(), currentSession.getUser().getUserId());
+  }
+
+  @Test
+  void performDeletionWorkflow_Should_commitDeletionBeforeErrorNotificationFails() {
+    prepareCurrentSessionForDeletion();
+    when(matrixSynapseService.deactivateUser(anyString())).thenReturn(false);
+    doThrow(new IllegalStateException("tenant unavailable"))
+        .when(workflowErrorMailService)
+        .buildAndSendErrorMail(anyList());
+
+    assertThrows(
+        IllegalStateException.class, deleteUserAnonymousScheduler::performDeletionWorkflow);
 
     assertSessionAndUserDoNotExistInDatabase(
         currentSession.getId(), currentSession.getUser().getUserId());
